Migration of static page menu and optimization of home page animation.
